refactor: finishes rationalizing components, renames and moves files
